refactor/parsec: make internal Observation crate visible only
Note:
The public our_unpolled_observations that is used by routing will now
construct BlockPayload for each item so it should not be used for
inspection.

elder/mod.rs is_trusted will need to be implemented using a local
cache instead (which will also be a lot more efficient).

clippy and test pass